+++ docs/source/version_history.rst
@@ -11,7 +11,8 @@ Bug Fixes
 * Fix GDML matrix loading for material properties where loading would create duplicate Expressions
   that would ultimately lead to conflicting definitions being written out. The written GDML file
   wouldn't be loadable in Geant4.
-* Fix geant4.Registry clear() function which had a variable error in it.
+* Fix geant4.Registry.clear() function which had a variable error in it.
+* Fix geant4.Registry.findPhysicalVolumeByName function.
 
 V1.0.5 - 2023 / 03 / 03
 =======================